Open_Source/3.Share_Things_With_Community/currency-converter-app-main/requirements.txt
streamlit==0.68.1
pandas==0.25.3
lxml==4.5.1
requests==2.22.0
json5==0.9.4
Pillow==7.0.0
